Clean Swift shell legacy fixtures
