+++ clients/apple/scripts/test-shell-sidebar-tab-rows.sh
@@ -16,6 +16,7 @@ CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/ShellStateMutations.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/ShellModel.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Models/Shell/ShellActionRegistry.swift" \
+    "$REPO_ROOT/clients/apple/scripts/support/ShellActionRegistryParitySupport.swift" \
     "$REPO_ROOT/clients/apple/scripts/test-shell-sidebar-tab-rows.swift" \
     -o "$TEST_BINARY"
